Host rewrite header (#19)
grpc issue: grpc/grpc#38062

I spent a while trying to figure out whether or not ServerName should take precedence over HostRewriteHeader. It seems that in the simplest config of a dns resolver, ServerName is being set, even though there are docs that say it should usually be empty.

Similarly when targeting an xDS target, ServerName is also set. So in order for this to work HostRewriteHeader must take precedence over ServerName.

Credit to Rob for the initial version of this.